[SourceKit] Hide `RenameRangeDetailCollector` as an implementation detail
The usage becomes a lot clearer if there’s just a top-level function that returns the details for the syntactic rename ranges.
